Pendenz 469

+++ athletica/lib/cl_result.lib.php
@@ -32,7 +32,8 @@ class Result
 	var $performance;
 	var $info;
 	var $points;
-  
+    var $remark;      
+    var $xAthlete;
 
 	function Result($round=0, $startID=0, $resultID=0)
 	{
@@ -40,11 +41,15 @@ class Result
 		$this->startID = $startID;
 		$this->resultID = $resultID;
 		$this->performance = '';
-		$this->info = '';                
+		$this->info = ''; 
+        $this->xAthlete = '';               
 	}
 	
-	function save($performance, $info = '', $secFlag = false)
-	{  
+	function save($performance, $info = '', $secFlag = false, $remark='', $xAthlete)
+	{   $this->remark = $remark; 
+        $this->performance = $performance; 
+        $this->xAthlete = $xAthlete;    
+     
 		require('./lib/utils.lib.php');
 		$GLOBALS['AA_ERROR'] = '';
 
@@ -65,16 +70,29 @@ class Result
 			return;
 		}
 
-		$performance = trim($performance);
-		
-		// Delete result (no performance entered)
-		if($performance == '')
-		{
-			$reply = $this->delete();
-		}
-		// Add or change result
-		else	
-		{
+         if (!is_null($performance)) {
+		    $performance = trim($performance);
+         }
+        
+        if (is_null($this->remark) && $performance == ''){  
+            $reply = $this->delete();  
+        }   
+        else
+            {            
+                if (is_null($performance) &&  $this->remark == ''){   
+                       $reply = $this->update_remark();     
+                }
+            
+                else {
+         
+                        if ( $this->remark != '' ){   
+                            $reply = $this->update_remark();     
+                        }
+            
+     
+   
+                        else
+		                {                
 			// validate performance
 			$retValidate = 0;
 			if($secFlag){
@@ -122,7 +140,8 @@ class Result
 				$reply = $this->update();
 			}
 		}
-		
+         }
+            }
 		return $reply;
 	}
 
@@ -196,6 +215,9 @@ class Result
 		$query = '';
 		$reply = new ResultReturn;
 
+        if ($this->performance!=''){
+           
+        
 		AA_utils_changeRoundStatus($this->round, $GLOBALS['cfgRoundStatus']['results_in_progress']);
 		if(!empty($GLOBALS['AA_ERROR'])) {
 			return;
@@ -216,7 +238,7 @@ class Result
 			$reply->setAction(RES_ACT_DELETE);
 		}
 		mysql_query("UNLOCK TABLES");
-
+        }
 		return $reply;
 	}
 
@@ -226,6 +248,100 @@ class Result
 		require('./lib/utils.lib.php');
 		$this->points = AA_utils_calcPoints($event, $this->performance, 0, $sex, $startID);
 	}
+    
+    function update_remark()
+    {   
+        $GLOBALS['AA_ERROR'] = '';
+        $query = '';
+        $reply = new ResultReturn();
+
+        mysql_query("
+            LOCK TABLES rundenset READ, runde READ, runde as r READ , serie as s READ , start as st READ, 
+            wettkampf as w READ , anmeldung as a READ , athlet as at READ, verein as v READ, 
+            rundentyp as rt READ, serienstart as ss READ  , serienstart WRITE
+        ");
+
+        if(!empty($this->startID))    // result provided -> change it
+        {
+            if(AA_utils_checkReference("serienstart", "xSerienstart"
+                                        , $this->startID) == 0)
+            {
+                $GLOBALS['AA_ERROR'] = $GLOBALS['strErrAthleteNotInHeat'];
+            }
+            else
+            {
+                $query="SELECT 
+                        w.mehrkampfcode , ss.Bemerkung
+                    FROM
+                        serienstart as ss
+                        LEFT JOIN start as st On (ss.xStart = st.xStart)
+                        LEFT JOIN wettkampf as w On (w.xWettkampf = st.xWettkampf) 
+                    WHERE
+                        w.mehrkampfcode = 0
+                        AND ss.xSerienstart = ".$this->startID;
+                
+                $result=mysql_query($query); 
+                
+                if(mysql_errno() > 0) {
+                    $GLOBALS['AA_ERROR'] = mysql_errno() . ": " . mysql_error();
+                }
+                else {
+                    if (mysql_num_rows($result) > 0) {
+                    
+                         $sql = "UPDATE serienstart 
+                                    SET Bemerkung = '".$this->remark."'                                     
+                                         WHERE xSerienstart = ".$this->startID.";";
+                            mysql_query($sql);
+                           
+                            if(mysql_errno() > 0) {
+                                $GLOBALS['AA_ERROR'] = mysql_errno() . ": " . mysql_error();
+                            }  
+                    }  
+                    else {
+                
+                      // comnined event
+                
+                        $query_mk="SELECT 
+                                    ss.xSerienstart , ss.Bemerkung   
+                               FROM 
+                                    runde AS r 
+                                    LEFT JOIN serie AS s ON (s.xRunde = r.xRunde) 
+                                    LEFT JOIN serienstart AS ss ON (ss.xSerie = s.xSerie)
+                                    LEFT JOIN START AS st ON (st.xStart = ss.xStart) 
+                                    LEFT JOIN wettkampf as w ON (w.xWettkampf = st.xWettkampf)
+                                    LEFT JOIN anmeldung AS a ON (a.xAnmeldung = st.xAnmeldung)
+                                    LEFT JOIN athlet AS at ON (at.xAthlet = a.xAthlet)
+                                    LEFT JOIN verein AS v ON (v.xVerein = at.xVerein   )
+                                    LEFT JOIN rundentyp AS rt ON rt.xRundentyp = r.xRundentyp                              
+                               WHERE 
+                                    w.mehrkampfcode > 0
+                                    AND at.xAthlet = ". $this->xAthlete;
+                           
+                        $result=mysql_query($query_mk); 
+                
+                        if(mysql_errno() > 0) {
+                            $GLOBALS['AA_ERROR'] = mysql_errno() . ": " . mysql_error();
+                        }
+                        else {
+                            while ($row=mysql_fetch_row($result)){
+                                    $sql = "UPDATE serienstart 
+                                            SET Bemerkung = '".$this->remark."'                                     
+                                            WHERE xSerienstart = ".$row[0].";";
+                                    mysql_query($sql);
+                           
+                                    if(mysql_errno() > 0) {
+                                        $GLOBALS['AA_ERROR'] = mysql_errno() . ": " . mysql_error();
+                                    } 
+                            }
+                        } 
+                    }
+                }
+            }
+        }
+        mysql_query("UNLOCK TABLES");
+
+        return $reply;
+    }
 
 } // end class Result
 
@@ -420,7 +536,7 @@ class ResultReturn
 	}
 
 } // end class ResultReturn
-
+    
 } // end AA_CL_RESULT_LIB_INCLUDED
 
 ?>
